+++ client/src/components/idea/discussion/CommentIcon.js
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import {AppContext} from "context";
 import React, {useContext} from "react";
-import {FaComment, FaCommentSlash, FaEdit, FaLink, FaLock, FaLockOpen, FaTags, FaUnlink, FaUserCheck} from "react-icons/all";
+import {FaComment, FaCommentSlash, FaEdit, FaLink, FaLock, FaLockOpen, FaTags, FaUndoAlt, FaUnlink, FaUserCheck} from "react-icons/all";
 
 const IconOverlay = styled.div`
   text-align: center;
@@ -45,6 +45,8 @@ const CommentIcon = ({specialType}) => {
                 return <Icon as={FaUnlink} fill={fill}/>;
             case "IDEA_ASSIGNED":
                 return <Icon as={FaUserCheck} fill={fill}/>;
+            case "IDEA_VOTES_RESET":
+                return <Icon as={FaUndoAlt} fill={fill}/>;
             case "LEGACY":
             case "TAGS_MANAGED":
             default: